+++ src/Persistence/Game/IPlayer.cs
@@ -31,6 +31,9 @@ namespace CivOne.Persistence.Game
 		ushort FutureTechCount { get; }
 		ushort HumanContactTurn { get; }
 		short StartX { get; }
+		(short X, short Y)[] MapPositions { get; }
+		string[] MapPositionNames { get; }
+		(short X, short Y) LastMapPosition { get; }
 		ushort[] UnitsLost { get; }
 		ushort[] UnitsDestroyedBy { get; }
 		ushort EpicRanking { get; }